requirements.txt
Django~=5.0.6
djangorestframework~=3.15.2
celery~=5.4.0
dj-config-url~=0.1.1